+++ packages/react-native-messaging-core/ios/Sources/NablaMessagingClient/NablaMessagingClientModule.m
@@ -9,6 +9,10 @@ RCT_EXTERN_METHOD(createConversation: (NSString *)title
                          providerIds: (NSArray<NSString *> *)providerIds
                             callback: (RCTResponseSenderBlock)callback)
 
+RCT_EXTERN_METHOD(createDraftConversation: (NSString *)title
+                              providerIds: (NSArray<NSString *> *)providerIds
+                                 callback: (RCTResponseSenderBlock)callback)
+
 RCT_EXTERN_METHOD(sendMessage: (id)input
                conversationId: (id)conversationIdMap
                       replyTo: (id)replyToMap